requirements.txt
fastapi
uvicorn[standard]
pymongo
pytest
python-dotenv
pydantic
ruff
python-jose[cryptography]
passlib[bcrypt]
httpx
python-multipart
